Add CI checks for Lighthouse & Bundle Size
Refactored CI into 3 workflows.

Lint & Test

--> Lint JS
--> Lint Handlebars (lol)
--> Test with coverage

Bundle Size

--> Extract JSON result from ember-cli-bundle-analyzer (scripts/bundle-size.js)
--> Upload results as workflow artifact
--> If it's a PR, compare the differences and comment them

Lighthouse

--> Run lighthouse and save scores
--> Upload results as a workflow artifact
--> If it's a PR, compare the differences and comment them

Additionally, added a couple dependencies to do our fragile little custom scripting.

Honestly...like...ok so here's how we check the bundle size. First, we start the server with ember s -prod. Then, we wait for it to be available, and run selenium with ChromeDriver to get the results of visiting the _analyze page. Once the title of the page changes, indicating the data is ready, we trawl through with htmlparser2 until we hit a <script> tag that has the data we're looking for. Since the data is stored as one massive javascript object, we can just yank it right out of the HTML and JSON.parse it.

Like...what the hell is that. How is that the easiest solution.

Anyway. More details in the PR.